transference feature working with converter implementation. everythig is working and persist in db

+++ ms-user/src/main/java/com/msuser/msuser/dto/DepositDTO.java
@@ -4,6 +4,8 @@ package com.msuser.msuser.dto;
 import lombok.*;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 /*Esta clase encapsula los datos que entran a través de la peticion de depositar dinero
 en el controlador (endpoint) para enviarlos con Rabbit al ms-transactions donde se procesará
 la petición */
@@ -12,11 +14,11 @@ la petición */
 @AllArgsConstructor
 @NoArgsConstructor
 public class DepositDTO {
+
     private String userId;
     private String accountNumber;
-    private float valueToTransfer;
+    private BigDecimal valueToTransfer;
     private String shippingCurrency;
     private String email;
-
 
 }
